rename TextareasPage to TextAreaPage
