+++ jdi-light-vuetify-tests/src/test/java/io/github/epam/vuetify/tests/complex/TextAreasTests.java
@@ -1,6 +1,5 @@
 package io.github.epam.vuetify.tests.complex;
 
-import com.epam.jdi.light.vuetify.elements.complex.TextArea;
 import io.github.com.dataproviders.TextAreasDataProviders;
 import io.github.com.entities.textareas.SignUpData;
 import io.github.epam.TestsInit;
@@ -8,25 +7,18 @@ import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 import static com.jdiai.tools.Timer.waitCondition;
-import static io.github.com.enums.Colors.BLACK_TRANSPARENT_087;
 import static io.github.com.enums.Colors.RED_ACCENT_2;
 import static io.github.com.StaticSite.textareasPage;
 import static io.github.com.custom.forms.textareas.SignUpForm.submit;
-import static io.github.com.pages.TextareasPage.appendInnerTextArea;
-import static io.github.com.pages.TextareasPage.appendOuterTextArea;
-import static io.github.com.pages.TextareasPage.autoGrowTextArea;
-import static io.github.com.pages.TextareasPage.autocompleteTextArea;
-import static io.github.com.pages.TextareasPage.clearableTextArea;
-import static io.github.com.pages.TextareasPage.counterTextArea;
-import static io.github.com.pages.TextareasPage.fourRowsTextArea;
-import static io.github.com.pages.TextareasPage.noResizeTextArea;
-import static io.github.com.pages.TextareasPage.oneRowTextArea;
-import static io.github.com.pages.TextareasPage.prependInnerTextArea;
-import static io.github.com.pages.TextareasPage.prependOuterTextArea;
-import static io.github.com.pages.TextareasPage.signUpForm;
-import static io.github.com.pages.TextareasPage.threeRowsTextArea;
-import static io.github.com.pages.TextareasPage.twoRowsTextArea;
-import static org.hamcrest.Matchers.containsString;
+import static io.github.com.pages.TextAreaPage.autoGrowTextArea;
+import static io.github.com.pages.TextAreaPage.clearableTextArea;
+import static io.github.com.pages.TextAreaPage.counterTextArea;
+import static io.github.com.pages.TextAreaPage.fourRowsTextArea;
+import static io.github.com.pages.TextAreaPage.noResizeTextArea;
+import static io.github.com.pages.TextAreaPage.oneRowTextArea;
+import static io.github.com.pages.TextAreaPage.signUpForm;
+import static io.github.com.pages.TextAreaPage.threeRowsTextArea;
+import static io.github.com.pages.TextAreaPage.twoRowsTextArea;
 import static org.hamcrest.Matchers.emptyString;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
